add fixtures file
