+++ interlecture/engine/js/main.jsx
@@ -2,35 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import Chat from 'js-common/chat';
 import Classroom from 'js-common/classroom';
-
-const classroom = {
-  id: 0,
-  lecturer: 'Kristian',
-  title: 'test',
-};
-
-const questions = [
-  {
-    id: 1,
-    user: 'Rolf',
-    body: 'hvorfor gjør vi dette?',
-  },
-  {
-    id: 2,
-    user: 'Arne',
-    body: 'kommer dette på eksamen?',
-  },
-  {
-    id: 3,
-    user: 'Kåre',
-    body: 'hva er meningen med livet?',
-  },
-  {
-    id: 4,
-    user: 'Bjarne',
-    body: 'finnes egentlig julenissen?',
-  },
-];
+import { classroom, questions } from './fixtures';
 
 const Main = () => (
   <div style={{ margin: '5pt 5%' }}>
